Extract ProcessRunsBuilder
